Added frontend UI

+++ backend/controllers/goalController.js
@@ -1,11 +1,13 @@
 const asyncHandler  = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 //@description: Get Goal
 //@route:       GET  api/goals
 //@access:      Private
 const getGoal = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
+
     res.status(200).json(goals)
 })
 
@@ -18,7 +20,8 @@ const setGoal = asyncHandler(async (req, res) => {
         throw new Error('Please add a text feild')
     }
     const goal = await Goal.create({
-        text:req.body.text
+        text:req.body.text,
+        user:req.user.id
     })
     res.status(200).json(goal)
 })
@@ -31,6 +34,19 @@ const updateGoal = asyncHandler(async (req, res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+    // Check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //Make sure the logged in user matches the goal's user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized!')
     }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new:true})
     res.status(200).json(updateGoal)
@@ -45,6 +61,19 @@ const deleteGoal = asyncHandler(async (req, res)=> {
         res.status(400)
         throw new Error('Goal not found')
     }
+    const user = await User.findById(req.user.id)
+    // Check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //Make sure the logged in user matches the goal's user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized!')
+    }
+
     await goal.deleteOne()
     res.status(200).json({id: req.params.id})
 })
